integrated chat new design and added notifications for unread messages
