+++ package.json
@@ -56,6 +56,7 @@
     "vue-instantsearch": "^3.7.0",
     "vue-loading-overlay": "^3.4.2",
     "vue-select": "^3.11.2",
+    "vue-simple-search-dropdown": "^1.0.1",
     "vue2-timepicker": "^1.1.6",
     "vuejs-datepicker": "^1.6.2",
     "vuelidate": "^0.7.6"
